homework7
