+++ src/context/firebaseContext.js
@@ -1,0 +1,3 @@
+import React from "react";
+
+export const firebaseContext = React.createContext(null)
