+++ src/routs/Finish/index.js
@@ -1,0 +1,9 @@
+const Finish = ()=>{
+    return(
+        <div>
+            <h1>Finish Page</h1>
+        </div>
+    )
+}
+
+export default Finish;
